split xarray.py into backend.py and accessor.py
